Statically manage extra files in SDK
We need to add some extra files to the SDK to make it functional, e.g.
/etc/make.conf. Rather than creating them at runtime in
build_sdk/setup.sh, commit them to the repository and use rules_pkg to
pass a bundled tarball to build_sdk. This will make it easy to add more
extra files to the SDK.

Change-Id: I39f399468b503b67aceae0449ca7f10a10cc4e02
Reviewed-on: https://team-review.git.corp.google.com/c/cros-build-tiger/cros-bazel/+/1573556
Reviewed-by: Raul Rangel <rrangel@google.com>