Use --skip-manifest
We don't want to generate digests, but instead we want to pretend they
don't exist. Bazel performs all the SHA verification on the downloaded
files already.

ebuilds like sys-firmware/fwupd-storage have a conditional SRC_URI.
We don't currently have these files listed as distfiles in the
BUILD.bazel because we don't need them right now. This package was
failing to build because `ebuild` was trying to generate the digest
for the missing files. This failed because it couldn't fetch the files.

By using --skip-manifest, `ebuild` will skip all manifest related
operations.

TEST=bazel build //third_party/chromiumos-overlay/sys-firmware/fwupd-storage \
     //third_party/portage-stable/sys-apps/attr

Change-Id: I48749740d30e33446cee271d7881f1db2acd6a15
Reviewed-on: https://team-review.git.corp.google.com/c/cros-build-tiger/cros-bazel/+/1582732
Reviewed-by: Shuhei Takahashi <nya@google.com>