ebuild_src: Allow specifying files in sub-packages
The chromium/src/third_party/shell-encryption package has a BUILD file
in the src directory. This prevents the standard `glob(["*"])` from
including the src directory in the `ebuild_src` target. This change
adds support for adding files in sub-packages. Without this, the
path for the sub-package files would be wrong.

Change-Id: I84d062eba1cef4679c9868952e6f9fc9877d647b
Reviewed-on: https://team-review.git.corp.google.com/c/cros-build-tiger/cros-bazel/+/1582733
Reviewed-by: Shuhei Takahashi <nya@google.com>