Copy all DEPEND deps into RDEPEND for rust src packages
rust src packages create a binary package that only has src code. In
order to create the binary package, the cros-rust eclass will install
the src into a cargo registry. This installation checks that all
dependencies exist in the registry, so we need to specify all the
deps in DEPEND.

Since src deps depend on other src deps, we want to pull in the full
dependency tree when actually compiling a rust package. For this reason
we need the src deps to also be listed in the RDEPEND field.

See https://bugs.chromium.org/p/chromium/issues/detail?id=1182669
for a similar issue. It looks like we have quite a few packages that
have been added since that fail to correctly specify their deps.

In order to detect rust src packages, I had to expose the eclasses used
and the presence of a src_compile function from the ebuild.

Now that we have the inherit information, we could in theory split up
the eclasses into different packages :)

Change-Id: I3434a36d117139cd5ba7394a9a66eef3bd3f933a
Reviewed-on: https://team-review.git.corp.google.com/c/cros-build-tiger/cros-bazel/+/1583936
Reviewed-by: Shuhei Takahashi <nya@google.com>