Generate ebuild targets for all rust packages
A lot of the rust packages contain multiple versions in different slots.
Some packages depend on one version, and others depend on another. This
means we can't just pick the ebuild with the greatest version, but
instead we need to generate targets for all the different versions.

This CL is kind of a hack. It's using the same dependencies and
distfiles for each version. This doesn't seem to be causing a problem,
but it is something we can improve upon.

In order to keep the dependencies simple I use a pacakge_set to group
all the versions together. This does mean that we end up installing
all the versions which isn't optimal, but it keeps us from having to
resolve the correct version number.

Ideally the whole rust tooling gets rewritten in bazel and we no longer
need to have a portage package for each crate version.

Rust builds are very slow because we need to install all the transitive
deps just to install a package into cargo and generate the binary
package.

TEST=bazel build //third_party/chromiumos-overlay/dev-rust/crossbeam-epoch

Change-Id: Ic67b4d72a07a48ee073182484ff1ecdbfa7b65e0
Reviewed-on: https://team-review.git.corp.google.com/c/cros-build-tiger/cros-bazel/+/1585694
Reviewed-by: Shuhei Takahashi <nya@google.com>