run_in_container: Replace --{diff,work}-dir with --scratch-dir
--diff-dir and --work-dir must be on the same filesystem otherwise
overlayfs won't work correctly. By having two flags it makes it possible
to pass in an invalid configuration. run_in_container also passes in the
diff dir and work dir unmodified to the overlayfs mount options. This
causes a problem if the file paths are long since the mount option
length is very constrained. By taking in a single directory we can
guarantee the work dir and diff dir are on the same mount, and we can
also guarantee we don't overflow the mount options since we now use
a relative path for the diff dir and work dir.

Change-Id: I1f8d765b1dae51c917d3512ffe56af0f96ce0649
Reviewed-on: https://team-review.git.corp.google.com/c/cros-build-tiger/cros-bazel/+/1596738
Reviewed-by: Shuhei Takahashi <nya@google.com>