Switch to using bind mount for sdk base
We can use the http_archive repo rule to unpack a tarball into a new
workspace. By directly using the root directory from the sdk workspace
we can avoid creating the squashfs, and instead just bind mount the
directory into the container. This means we no longer need to use
squashfuse for the sdk base, and we get proper OS file caching. This
drops the kinux kernel build time on my workstation down from 500s to
300s.

This change in theory still keeps the build hermetic. The sdk workspace
is managed by bazel. When ever any properties of the workspace change
bazel will delete it and recreate it from scratch. To be extra cautious
I create a new repository for each SDK. This means that we will get a
different path for each base SDK, and it should cause a cache bust when
the SDK version is changed. I imagine we can write a script to generate
sdk/repositories.bzl in the future by scraping the gs bucket.

I tested this by building a package, changing the sdk version, built
the package again, and verified it was rebuilt.

Future improvements would be to stop using http_archive and instead
write out own repo rule. We could have this repo rule call `pixz`
to extract the tarball since bazel takes about 5 minutes to extract
the sdk tarball.

I plan on seeing if I can convert the sdk_pkgs.squashfs into a bind
mount directory as well.

Change-Id: Ieec133cd3ea44e4b04423095dcd8737a4c8e82a1
Reviewed-on: https://team-review.git.corp.google.com/c/cros-build-tiger/cros-bazel/+/1596741
Reviewed-by: Shuhei Takahashi <nya@google.com>