sdk: Use tar or pixz to extract the base sdk
Apparently download_and_extract mangles the symlinks. It appends the
execRoot to all absolute symlinks. This is obviously problematic because
they now point to an invalid location. If we use tar to extract the
tarball, the symlinks are left intact. I also added an option to use
pixz if it's available on the host. With pixz extracting the base sdk
drops from ~5 minutes to ~1 minute.

Change-Id: I50fd584ee253d53305d5a59eebfe111b3743b628
Reviewed-on: https://team-review.git.corp.google.com/c/cros-build-tiger/cros-bazel/+/1597614
Reviewed-by: Shuhei Takahashi <nya@google.com>