extract_deps: Parse SRC_URI and return it in deps.json
This will allow us to trim down the set of distfiles we pass in to each
ebuild. It also passes along the manifest values so `update_build`
doesn't have to parse the Manifest.

I didn't update `update_build` in this CL to read this data because
`chromeos-base/lacros` and some of the `dev-lang/rust` ebuilds
are relying on the fact that we pass in all the dist files. The `lacros`
fix is easy, we just need to parse the non 9999 ebuild to get the
correct SRC_URI. We already exclude the 9999 version in `update_build`,
we just need to apply the same hack to `extract_deps`. The `rust`
ebuilds will require a larger refactor. Since we support multiple
versions of the packages, we need to update `extract_deps` to support
multiple versions. This way we can get the correct SRC_URIs for the
correct versions. This will also fix the problem where we apply the
same deps to all package versions. This is causing problem because
they aren't always the same.

Change-Id: I45418c8a51a95a0dd06c012d17fc6424f4435478
Reviewed-on: https://team-review.git.corp.google.com/c/cros-build-tiger/cros-bazel/+/1601893
Reviewed-by: Shuhei Takahashi <nya@google.com>