Fix circular dependency caused by rust packages
In order to fix this we make extract_deps return the specific version
of the rust package that we need. To do this I converted the buildDeps
and runTimeDeps to contain a package atom instead of a simple package
name.

I kept this CL as a different CL to make it easier to review, but
I can also squash it with the parent.

Change-Id: I5a0b9177456da7544395d07f91607ef88a2ea794
Reviewed-on: https://team-review.git.corp.google.com/c/cros-build-tiger/cros-bazel/+/1603318
Reviewed-by: Shuhei Takahashi <nya@google.com>