Install packages in parallel
When I previously tried to pass --jobs to emerge it failed
spectacularly because we were also passing in --nodeps. When I got rid
of --nodeps emerge spend a lot of time calculating dependencies, so It
didn't seem worth it. Since bazel has the dep graph, I can construct
groups of deps to install in parallel. This results in a drastic
speed up when installing packages that have a lot of dependencies.
These packages are usually the long poll and would cause the build
to take over 4 hours!

TEST=bazel build //:all_target_packages

Change-Id: I33527c7da04cc11ff8efc4488166e776a19bbbfb
Reviewed-on: https://team-review.git.corp.google.com/c/cros-build-tiger/cros-bazel/+/1605733
Reviewed-by: Shuhei Takahashi <nya@google.com>