extract_deps: Add missing subdir src targets
Not all packages use the same level of SUBTREE, some have deeper targets
than others. This results in the packages that have a shallower SUBTREE
missing out on the files defined in the deeper tree.
To fix this we need to populate targets with the shallow tree with
all the additional deeper paths.

e.g.,
iioservice requires //platform2/iioservice:src
cros-camera-libs requires //platform2/iioservice/mojo:src

When trying to build iioservice the mojo directory will be missing.

Change-Id: Ie947bdaf5b8278ebb3d75db9a9df88e7dfba489e
Reviewed-on: https://team-review.git.corp.google.com/c/cros-build-tiger/cros-bazel/+/1607284
Reviewed-by: Shuhei Takahashi <nya@google.com>