build_package: Use . as tmp directory
The build_pacakge tmp dir is currently using /tmp. This makes it
difficult to locate the build artifacts. It also doesn't get cleaned
up when building the package fails. Create the build_package tmp dir in
the `execroot` to make it easy to find the artifacts and have bazel
clean up the directory.

Change-Id: I60806c938d2f3ccf353c3c06bf03097855c44736
Reviewed-on: https://team-review.git.corp.google.com/c/cros-build-tiger/cros-bazel/+/1615993
Reviewed-by: Shuhei Takahashi <nya@google.com>