Add a dependency on libfuse in e2fsprogs.
Apparently, while e2fsprogs doesn't support the --with-fuse flag, rules_foreign_cc automatically adds -L$EXT_BUILD_DEPS/fuse/lib and -I$EXT_BUILD_DEPS/fuse/include, so it just magically works.

Using ldd, it appears that the correct .so file is used, and looking at the invocation of make, it appears to be invoked with the correct environment variables.

Change-Id: I788d2ce96a0b6a1308c39a3072f29766251e793b
Reviewed-on: https://team-review.git.corp.google.com/c/cros-build-tiger/cros-bazel/+/1615482
Reviewed-by: Shuhei Takahashi <nya@google.com>