Use specific version of bazel
This pulls in bazelisk as a host dependency, but it allows us to easily
run different versions of bazel. I chose the latest pre-release version
so we can test it and file any bugs.

I also included the missing command to install squashfs-tools. In the
future we should build this from source to not rely on the host version.
Or even better, contribute a bzlmod entry for it :)

Change-Id: Ib734b695cd4f620bf5a54992d7d9c9430954aae0
Reviewed-on: https://team-review.git.corp.google.com/c/cros-build-tiger/cros-bazel/+/1621996
Reviewed-by: Matt Stark <msta@google.com>
Reviewed-by: Shuhei Takahashi <nya@google.com>