Install ebuild in proper overlay and include .bashrc
samba and brltty both have .bashrc files next to their ebuild files. We
need to add these to the glob so they get included in the build. These
.bashrc files also only get used if the package is part of the
chromiumos-overlay. So we need to install the ebuild in the proper
overlay directory instead of randomly picking the first one.

Instead of passing in an `--overlay` flag like we do for `--category`,
I just parse the ebuild path. This lets us remove the `--category` flag
as well.

Change-Id: Ic82bbb470e230d571f1b35642a42b4e33cef47b0
Reviewed-on: https://team-review.git.corp.google.com/c/cros-build-tiger/cros-bazel/+/1626452
Reviewed-by: Shuhei Takahashi <nya@google.com>