Support globbing sub-packages
We can use the --deleted_packages command to tell bazel to ignore
parsing a BUILD file. This allows `glob` to correctly grab all the src
for a package. Ideally we would have a .bazelrc in each project that
needed to ignore these BUILD files, but we can only specify the
--deleted_packages arg once, so we need a global list.

Change-Id: Ibdcbe797d579bf081e192efd1a31eab3af8433c6
Reviewed-on: https://team-review.git.corp.google.com/c/cros-build-tiger/cros-bazel/+/1629363
Reviewed-by: Shuhei Takahashi <nya@google.com>