sdk/usr/local/bin: Fix incorrect nesting
The :bin target had a prefix of "usr/bin" and the :files target had a
prefix of usr/local/bin, this resulted in the files getting placed in
/usr/bin/usr/local/bin which is incorrect. This change reverts back to
the original version before fakefs was introduced. fakefs now gets
installed into /usr/bin/fakeroot, so we no longer need it here.

Change-Id: I32a737df7f87b894808dcb358eb9f65f5aaeeb77
Reviewed-on: https://team-review.git.corp.google.com/c/cros-build-tiger/cros-bazel/+/1634193
Reviewed-by: Shuhei Takahashi <nya@google.com>