Add a repository rule to fetch chrome
This CL adds a new repository rule to download chromium, download the
sub-modules via gclient, populate cipd, run the gclient hooks, and
generate a tarball. I chose zstd for the tarball since it should be
installed by default, and it's way faster than gzip. We could in theory
add tar and zstd into cipd to have a stable version.

We then wrap the tarball in an ebuild_tar_src rule so we can specify
the mount point. The chrome ebuild uses /home/$USER/chrome_root by
default, so I chose to stick with that. We could always override it
via the CHROME_ROOT env variable if necessary.

I chose to treat the tarbal as a src file instead of a distfile because
distfiles are copied into the container, and extracted in the container.
This means we would be copying 6 GiB into the container, then extracting
it. By treating it as a src we can do the extracting while creating
the container and thus avoid the expensive copy.

Change-Id: I36175e2838c1329b599409d77e6bab6ba5a683d2
Reviewed-on: https://team-review.git.corp.google.com/c/cros-build-tiger/cros-bazel/+/1632392
Reviewed-by: Shuhei Takahashi <nya@google.com>