perf: Use ld.so.preload to speed up most stat(2) syscalls
This change introduces /usr/lib64/libfakefs_preload.so to be preloaded
to the dynamic loader via /etc/ld.so.preload to hook glibc stat* family
function calls. The hook functions check if the target file has the
fakefs xattr by calling {,l,f}getxattr. If the xattr is found, pass
through the call to glibc fstatat, which is eventually intercepted by
fakefs (if it is running). Otherwise (which is 99% of the case), call
newfstatat(2) with the "backdoor key" set on the unused 6th argument of
the syscall.  fakefs' seccomp-BPF filter is updated to pass through
syscalls if the 6th argument matches the backdoor key.

With this approach, we can reduce fakefs overhead by avoiding context
switches in many cases.

Total build time: 74m 36s -> 53m 49s

Note that I also tried setting LD_PRELOAD environment variable as well,
but it didn't work when Portage spawns subprocesses without inheriting
the environment variable. Thus we use /etc/ld.so.preload instead.

Change-Id: I7308d8d4cca72200b2e7ba33bbcc0029811709c1
Reviewed-on: https://team-review.git.corp.google.com/c/cros-build-tiger/cros-bazel/+/1632719
Reviewed-by: Raul Rangel <rrangel@google.com>