fix: Avoid directly calling os.Exit to avoid leaking deferred calls
os.Exit terminates the current process immediately without calling
defered function calls. This leads to leaking resources, such as leaving
temporary directories, if we call os.Exit in the middle of a program.

This patch introduces "cliutil" package that provides utilities to
avoid this. When a program wants to exit with a certain exit code, it
can return cliutil.ExitCode as an error. It will be eventually processed
by the top-level main function that calls cliutil.Exit that in turn
calls os.Exit with the specified exit code.

Change-Id: I2a7604b7fe5770ff30f76fbdb186d52509ee0340
Reviewed-on: https://team-review.git.corp.google.com/c/cros-build-tiger/cros-bazel/+/1637695
Reviewed-by: Raul Rangel <rrangel@google.com>