run_in_container: Remove RUNFILES_DIR and RUNFILES_MANIFEST_FILE
These environment variables contain an absolute path on the build
machine. This absolute path contains the user name, and the bazel
workspace hash. When portage builds a binpkg, it snapshots the whole
environment. So we need to exclude build host specific data to get
a reproducible build.

Since the RUNFILES_DIR won't be valid after the pivot, let's clear the
variables before launching the executable. If we do need these variables
down the line, we can make build_packages strip them out instead.

Change-Id: I70e74a14e484c21fa3c99f440c42fd9567afb96b
Reviewed-on: https://team-review.git.corp.google.com/c/cros-build-tiger/cros-bazel/+/1640878
Reviewed-by: Shuhei Takahashi <nya@google.com>