run_in_container: Make bind mounting a bit smarter
This change will check if the target file exists, and if it doesn't then
if the source is a directory it will create a target directory,
otherwise it will create an empty file. This will allow directories to
be bind mounted, and in addition named pipes. The old code was always
calling os.WriteFile. This is problematic for named pipes, because it
will try and open the pipe and write to it. Checking for existence
solves this problem.

Change-Id: I2919bbbb8f405adac2969570e8bcd34c3006d02c
Reviewed-on: https://team-review.git.corp.google.com/c/cros-build-tiger/cros-bazel/+/1668652
Reviewed-by: Matt Stark <msta@google.com>
Reviewed-by: Shuhei Takahashi <nya@google.com>