mountsdk: Fix TPGID after debug bash exits
In order to restore the TPGID (Terminal Processes Group ID) after bash
exists, we need to have a processes outside the container restore the
TPGID because processes inside the container don't have access to the
original PGID. This is performed by passing in a control pipe into the
container and having a listener outside the container that will handle
the request.

With this change keyboard signal handling works correctly after bash
exits.

TEST=bazel run //third_party/portage-stable/sys-libs/zlib:0_debug -- --login=before
Then press Ctrl+C and watch ebuild action terminate.

Change-Id: Ic488a926a7ef16bfb68a0a7130530b9bb565872b
Reviewed-on: https://team-review.git.corp.google.com/c/cros-build-tiger/cros-bazel/+/1669718
Reviewed-by: Matt Stark <msta@google.com>
Reviewed-by: Shuhei Takahashi <nya@google.com>