processes: Add RunProcess which handles terminating child processes
The behavior of exec.CommandContext doesn't fit our use case. When the
context terminates, it sends a SIGKILL to the process. This means the
child process doesn't have time to clean up. In our case that means
we kill the `run_in_container` processes which started PID1. This leaves
PID1 orphaned and thus the container running. i.e., we are leaking
containers on Ctrl+C/SIGTERM.

This CL adds a new method RunProcess that will handle SIGTERM and SIGINT
correctly, and not leak any child processes.

Change-Id: I18a0b734e1827a092c5b0951492c63dc1b4d451d
Reviewed-on: https://team-review.git.corp.google.com/c/cros-build-tiger/cros-bazel/+/1670933
Reviewed-by: Shuhei Takahashi <nya@google.com>
Reviewed-by: Matt Stark <msta@google.com>