ebuild: Add ability to specify shared_lib_deps
When an ebuild specifies shared_lib_deps, the shared_libs and headers
specified in the dep will be included in the container. This avoids
having to depend on the whole tbz2 to build. It also opens the door
to injecting the `interface` .so's. This only handles 1 level of deps,
I haven't plumbed in the transitive part. I'll do that as a followup.

With this change I was able to build squashfs-tools using the
following changes to the ebuilds: https://paste.googleplex.com/6458886126043136

I think this proves the basic concept.

Change-Id: Icf67689a26f6ce8987925e072ecf1f8d56be9d7a
Reviewed-on: https://team-review.git.corp.google.com/c/cros-build-tiger/cros-bazel/+/1672382
Reviewed-by: Shuhei Takahashi <nya@google.com>