xpak: Add the extract-metadata sub-command
This sub-command parses the binpkg and generates a .ext file that lives
next to the .ebuild. This additional metadata will be parsed as part
of extract_deps and used when computing the dependency graph. The goal
is to trim the current dependency graph so we only pull in exactly what
is needed to build a package.

An example of the generated output can be found here:
https://team-review.git.corp.google.com/c/cros-build-tiger/third_party/chromiumos-overlay/+/1674244

TEST=bazel run //bazel/ebuild/cmd/xpak:xpak -- extract-metadata -w *.tbz2

Change-Id: I8351bc7681fcc5b2058b0c1294134f77c053baa2
Reviewed-on: https://team-review.git.corp.google.com/c/cros-build-tiger/cros-bazel/+/1674511
Reviewed-by: Shuhei Takahashi <nya@google.com>