Properly use three-valued logic in dependency predicates
Surprisingly, Portage uses three-valued logic on evaluating dependency
satisfaction. Notably, USE-conditional dependency returns "null"
instead of true when its USE flag precondition is not satisfied.
For example, `|| ( foo? ( a/b ) )` should be considered unsatisfiable
when `foo` is unset.

To implement this, this patch:
- Updates the Dependency type to return three-valued logic in matches()
  function.
- Introduces flat_map_tree/try_flat_map_tree methods that allows mapper
  functions to return Option<D>. They are used in elide_use_flags().
- Implements the Default trait for the Dependency type. Now that
  elide_use_flags() may return an empty dependency, it is convenient to
  be able to use unwrap_or_default() to treat it as the constant true
  dependency.

This change makes actual differences on analyzing package dependencies.
For example, virtual/libgudev package has the following RDEPEND:

  || (
    dev-libs/libgudev
    !systemd? ( ... )
    systemd? ( ... )
  )

If we consider unsatisfied USE conditionals as the constant true, this
dependency expression as a whole is simplified to the constant true
because either one of !systemd? ( ... ) or systemd? ( ... ) is the
constant true regardless of the value of systemd. But in fact, it should
be parsed as an any-of dependency between dev-libs/libgudev and other
candidates that differ by the value of systemd.

Change-Id: I81d51f57435674e87f78ec91deb9a1c7c08719bb
Reviewed-on: https://team-review.git.corp.google.com/c/cros-build-tiger/cros-bazel/+/1687943
Reviewed-by: Raul Rangel <rrangel@google.com>