refactor: Extract reusable analysis logic from dump_deps
dump_deps contains functions to analyze ebuild metadata and compute
dependencies and source information. This patch extracts them as
separate modules for reuse.

Previously, the "translate" module provided some functions for analysis,
but it was too small to be reusable. It is now merged to the new
"analysis" module.

This is a refactoring change. No functional change is expected.

Change-Id: If755cf6191bee116ee0e8a1cb86d3d2b428ca694
Reviewed-on: https://team-review.git.corp.google.com/c/cros-build-tiger/cros-bazel/+/1690257
Reviewed-by: Raul Rangel <rrangel@google.com>