Introduce a subcommand to generate a repository
This change adds a new subcommand "generate-repo" to alchemist,
which analyzes *all* ebuilds in overlays and generate a directory
containing a bunch of symlinks pointing to the original overlays and
BUILD.bazel files generated on the fly. This is done without going
through deps.json.

We still don't run the command to generate a repository automatically in
Bazel repository rules as it's not complete yet.

We use the version number (e.g. 2.5.1) to identify a package version,
instead of main SLOT (e.g. 0) we used to use with extract_deps. This is
because we want to generate Bazel targets for all ebuilds, instead of
those transitively depended by virtual/target-os*, so that users can
build packages not preinstalled to ChromeOS images. We may revisit this
later since a package version rebump may cause all transitive reverse
dependencies to be rebuilt.

TEST=cargo run --release -- --board=arm64-generic generate-repo

Change-Id: Ia88cff5bcc84a995cdfef34677185971941d8052
Reviewed-on: https://team-review.git.corp.google.com/c/cros-build-tiger/cros-bazel/+/1690846
Reviewed-by: Raul Rangel <rrangel@google.com>