refactor: Make Resolver focus on resolving packages
Previously, Resolver was a monolithic object that owns various other
objects, such as RepositorySet, ConfigBundle and EBuildEvaluator, that
were initialized in Resolver's constructor. This design was convenient
to implement extract_deps, but it's not reusable for other purposes
because various important objects were owned by Resolver and thus could
not all their interesting methods.

This patch renames Resolver to PackageResolver and make it focus on
resolving package atoms, and borrow various other objects instead of
owning them. Those objects will be initialized explicitly in the main
function.

Change-Id: I3b47eeecb2f4ad89f8d809119c7714bc90f2198f
Reviewed-on: https://team-review.git.corp.google.com/c/cros-build-tiger/cros-bazel/+/1690258
Reviewed-by: Raul Rangel <rrangel@google.com>