build_package: Pass inside and outside path for --ebuild
If the ebuild BUILD.bazel file lives outside of the main repository,
the path will have the workspace prefixed.

i.e., external/XXX/third_party/portage-stable/...

This causes the ebuild to be mounted in the wrong spot. To fix this we
pass in the path we want the ebuild mounted.

Change-Id: I7bfe066f61bc130425f3a6520ce8e96b4383f896
Reviewed-on: https://team-review.git.corp.google.com/c/cros-build-tiger/cros-bazel/+/1695060
Reviewed-by: Shuhei Takahashi <nya@google.com>