sdk: Add amd64-generic SDK
This CL renames the sdk target to arm64-generic and adds a new target
for amd64-generic. We generally only have 3 architectures so adding
another x86 board or arm64 board should be pretty straight forward. I
chose the $BOARD name as the target name to make it easy to associate
the generated BUILD files with the correct SDK. I'm hoping that we can
have alchemist parse the toolchain.conf file and generate a proper SDK
target. I'm also hopeful that we can replace
all the individual make.conf files with a single one that is generated
by alchemist.

This change keeps the update_build flow working for the arm64-generic
board. I will not add the ability to build different boards to
update_build since we will focus on alchemist instead.

TEST=bazel build @board//third_party/chromiumos-overlay/sys-libs/readline:all

Change-Id: I899596ffe78c2498019c36161eccab437607641b
Reviewed-on: https://team-review.git.corp.google.com/c/cros-build-tiger/cros-bazel/+/1695397
Reviewed-by: Shuhei Takahashi <nya@google.com>