Improve debuggability of ebuild analysis failures
The most popular reason for failing to analyze an ebuild is that its
dependency evaluates to the constant false due to missing packages.
This change tries to explain such failures by tracking the reason why
a dependency was deduced to the constant false.

Good message example:

WARNING: .../net-misc/chrony/chrony-4.2.ebuild: Analysis failed:
  Resolving build-time dependencies for net-misc/chrony-4.2:
  Unsatisfiable dependency: No package satisfies acct-group/ntp

We still fail to explain some cases nicely, e.g.

WARNING: .../app-benchmarks/glmark2/glmark2-285.ebuild: Analysis
  failed: Resolving build-time dependencies for
  app-benchmarks/glmark2-285: Unsatisfiable dependency: Unknown

This typically happens when all children of an all-of expression
are deduced to the constant false. But in most cases the message
is better.

TEST=cargo run --release -- --board=arm64-generic generate-repo

Change-Id: Ib9d442ac5a62cfbd1f26bfc19523276c72660fe5
Reviewed-on: https://team-review.git.corp.google.com/c/cros-build-tiger/cros-bazel/+/1696193
Reviewed-by: Raul Rangel <rrangel@google.com>