install_deps: Clean up layer
install_deps was not properly cleaning up the layer, so we were getting
non-hermetic outputs. This causes problems with remote caching.

sdk_update already has code to handle the non-hermetic outputs, so we
just need to add this to install_deps. Since we don't have an easy way
to share `bash` libraries between utils, I migrated the ~30 lines of
bash to ~130 of go code.

TEST=build a package twice and diff the deps layer

Change-Id: I0381cdc0f1d7ed5dd0a3010811b80bc2166aba41
Reviewed-on: https://team-review.git.corp.google.com/c/cros-build-tiger/cros-bazel/+/1699986
Reviewed-by: Shuhei Takahashi <nya@google.com>
Reviewed-by: Ryo Hashimoto <hashimoto@google.com>