sdk: Build board's primordial packages instead of using binpkgs
In order to do this we need to hard code a list of primordial packages
into alchemist. These packages can use the `base` board SDK that only
contains the tool chains. This is required to avoid a circular
dependency.

With this, we should now be able to generate the sdk targets in
alchemist.

TEST=bazel build @board//third_party/portage-stable/sys-libs/zlib:all

Change-Id: I7c6fc97e4be6423efb9b4d182a1ca1790ca7e90d
Reviewed-on: https://team-review.git.corp.google.com/c/cros-build-tiger/cros-bazel/+/1699988
Reviewed-by: Shuhei Takahashi <nya@google.com>