rust: Stop generating Cargo.toml
One of our goals is to compile alchemy from a repository rule. Alchemy
is currently using `workspace` dependencies. This means we need to have
the Cargo.toml present during the loading phase, meaning it needs to be
a real file.

This change deletes the Cargo.toml generation logic and replaces it with
static file. This means that we have two places to update when adding a
new cargo package, but I'm hoping we can reduce that down to one:
https://github.com/bazelbuild/rules_rust/issues/1199

This package also changes the non-root Cargo.toml files to use `path`
to declare local dependencies. This makes it clear that these are local
crates and not 3rd party crates. If we don't like having the paths
defined in the Cargo.toml files we can move them back to the root
Cargo.toml and delete the [members] section since using a `path` for a
dependency will automatically add the crate as a member.

BUG=b:265386129
TEST=`cargo build` and see all packages getting built.
In order to run `update_crates`, I needed to skip using
`update_crates_custom_bazel`. That was happening before this CL, so I
think it's unrelated.

Change-Id: Ib02a26ca74bc8ad72f7e77c1fcdaea056e094314
Reviewed-on: https://team-review.git.corp.google.com/c/cros-build-tiger/cros-bazel/+/1705498
Reviewed-by: Matt Stark <msta@google.com>
Reviewed-by: Shuhei Takahashi <nya@google.com>