Fix building packages now that they live in internal/layout
The ebuilds were getting placed in the wrong location inside the chroot.
If we didn't want to hard code this, we could add an `overlay_prefix`
to the ebuild rule and generate that, or have the `ebuild` point to an
`overlay_set`. For now I think stripping the prefix is a simple enough
solution.

TEST=bazel build  @board//sys-libs/zlib:all

Change-Id: Ibf6da4783f6aba072d065e81315a13a04184f812
Reviewed-on: https://team-review.git.corp.google.com/c/cros-build-tiger/cros-bazel/+/1706597
Reviewed-by: Shuhei Takahashi <nya@google.com>