Add @portage repository
This repository will invoke `alchemist generate-repo` using the BOARD
passed in from the @portage_cache repository. Since the repository rule
consumes the `digest` file, the repository rule will only re-run when
the digest changes. Right now the digest implemented as a TODO, so it's
always changing, so this will re-run the repository everytime it's
referenced.

This rule will only work correctly if the user has ran
`setup_board --board <board>` inside the `cros_sdk`. The next step is
to break the dependency on the `cros_sdk` and instead use the
//bazel/sdk:base_sdk and generate the required make.conf files that
setup_board creates.

BUG=b:263139238
TEST=BOARD=amd64-generic bazel query @portage//sys-libs/zlib:all
@portage//sys-libs/zlib:1.2.11-r6
@portage//sys-libs/zlib:1.2.11-r6_debug
@portage//sys-libs/zlib:1.2.11-r6_package_set

Change-Id: I0597239b379e82628c91d5279237c6890e9751cf
Reviewed-on: https://team-review.git.corp.google.com/c/cros-build-tiger/cros-bazel/+/1708592
Reviewed-by: Shuhei Takahashi <nya@google.com>