Add @portage_digest repository
This repository calls `alchemist digest-repo` to generate a digest
for the board's overlays. This digest will be used to cache bust the
@portage repository when anything changes. This repository will be
regenerated every single time the `bazel` command is invoked due to the
`_CACHE_BUST_DATE` environment variable that was added to tools/bazel.
For this reason we need to make sure `digest-repo` will be as fast as
possible.

BUG=b:263139238
TEST=BOARD=amd64-generic bazel build @portage_digest//:digest && cat bazel-src/external/portage_digest/digest

Change-Id: Ic862d76c2a6f7e1e739055e5d22bc466989a36b1
Reviewed-on: https://team-review.git.corp.google.com/c/cros-build-tiger/cros-bazel/+/1708591
Reviewed-by: Shuhei Takahashi <nya@google.com>