alchemist: Create a place holder digest-repo command
This command will be used to generate a digest of the overlays. This
digest will be used to trigger a rebuild of the @portage repository when
any of the content changes. I left it as a TODO so we can get the rest
of the scaffolding in place. It will currently cache bust every single
time.

BUG=b:263139238
TEST=cargo run -- --board amd64-generic digest-repo

Change-Id: I1f82c1cd0c63a05294b32963fe7aaab994a07a62
Reviewed-on: https://team-review.git.corp.google.com/c/cros-build-tiger/cros-bazel/+/1708590
Reviewed-by: Shuhei Takahashi <nya@google.com>