alchemist: Ensure PackageLocalSource corresponds to directories
CROS_WORKON_SUBTREE allows regular files to be specified. They are
not easy to handle on generating source packages.

This patch works around the problem by rewriting source specification
to a parent directory when it points to a regular file.

This doesn't affect the output of dump-deps, at least for arm64-generic.

BUG=b:262460684
TEST=cargo run --release -- --board=arm64-generic dump-deps \
     virtual/target-os{,-dev,-test} > ../../../data/deps.json

Change-Id: I343de3e3762e20ada0d7c8a3a4b101381825f65a
Reviewed-on: https://team-review.git.corp.google.com/c/cros-build-tiger/cros-bazel/+/1714543
Reviewed-by: Raul Rangel <rrangel@google.com>