alchemist: Break dependency on the chroot
The chroot/sdk actually provides minimal configuration. Most of the
config files just point into the chromiumos-overlay. By hard coding the
few files we can break the dependency on the chroot and having to run
`setup_board`. With this change we can generate arbitrary boards!

I had to do some refactoring on the path translator to support
bi-directional translation. Since we only have 1 bind mounted path I
also simplified the translator. I made it so alchemist generates the
config files because it has access to the generate_repo templates while
fakechroot doesn't. This makes fakechroot pretty generic so we could use
it for something else if necessary.

BUG=b:265214747
TEST=cargo run -- --board brya generate-repo --output-dir ~/cros-bazel/board-brya
TEST=cargo run --release -- --board=arm64-generic dump-deps virtual/target-os{,-dev,-test} > ../../../data/deps.json

Change-Id: Ida0d05cffa3094279f8cba5a99bd0f9516d0e229
Reviewed-on: https://team-review.git.corp.google.com/c/cros-build-tiger/cros-bazel/+/1717708
Reviewed-by: Shuhei Takahashi <nya@google.com>