Start a public bazel migration design document
My goal for this document is for it to serve as the source of truth on
how the bazel migration was implemented and to contain the reasoning
behind the decisions we made.

There is no individual owner, and I would like it if everyone worked
together to contribute and keep this document up to date.

BUG=none
TEST=none

Change-Id: Iff34afdd8952d3cc8f8fd229ff1df0dd8a58ed52
Reviewed-on: https://team-review.git.corp.google.com/c/cros-build-tiger/cros-bazel/+/1724542
Reviewed-by: Shuhei Takahashi <nya@google.com>