alchemist: Switch min stability to stable but allow cros-workon packages
Building all unstable packages by default can result in us trying to
build packages that don't actually function. This change switches the
min stability level to stable. This means all cros-workon 9999 packages
get excluded when computing the dependency graph. We don't actually want
to require the `cros_workon` command anymore, so we add an exception for
9999 cros-workon packages that don't have CROS_WORKON_MANUAL_UPREV set.
These are packages that normally get upreved by annealing so it's fine
to use the local src.

With this change and the previous change that allows building pinned
packages, we can now remove the pinned rust crosvm repositories in the
manifest.

 alias(
     name = "cros_async",
-    actual = ":9999",
+    actual = ":0.1.0-r41",
     visibility = ["//visibility:public"],
 )
 alias(
     name = "data_model",
-    actual = ":9999",
+    actual = ":0.1.0-r50",
     visibility = ["//visibility:public"],
 )
 alias(
     name = "qemu",
-    actual = ":9999",
+    actual = ":5.2.0-r51",
     visibility = ["//visibility:public"],
 )
 alias(
     name = "chrome-icu",
-    actual = ":9999",
+    actual = ":107.0.5257.0_rc-r1",
     visibility = ["//visibility:public"],
 )
 alias(
     name = "chromeos-chrome",
-    actual = ":9999",
+    actual = ":107.0.5257.0_rc-r1",
     visibility = ["//visibility:public"],
 )

The chrome packages are interesting. They rely on the @chrome repo, so
we need to generate the @chrome repo so that the version numbers match.


 alias(
-    name = "9999",
-    actual = "//internal/overlays/third_party/chromiumos-overlay/chromeos-base/chromeos-chrome:9999",
-    visibility = ["//visibility:public"],
-)

Analysis failed: Resolving build-time dependencies for chromeos-base/chromeos-chrome-9999: Unsatisfiable dependency: No package satisfies ~chromeos-base/chrome-icu-9999

This one is unfortunate. With this change we can no longer build a 9999
chromeos-chrome because we don't have a way of marking chrome-icu as
stable. I'm not sure how we would even go about building a 9999
chromeos-chrome since the repository rule needs to download the src
code. We might need to have a special flow for chrome edit/debug cycles.
For now I don't think it's a problem.

BUG=b:262460834
TEST=cargo run -- --board arm64-generic generate-repo --output-dir ~/tmp/board-stable-only/ --verbose

Change-Id: I6a95ee200336747581b200f9ac549007cf29c983
Reviewed-on: https://team-review.git.corp.google.com/c/cros-build-tiger/cros-bazel/+/1756963
Reviewed-by: Ryo Hashimoto <hashimoto@google.com>
Reviewed-by: Shuhei Takahashi <nya@google.com>