alchemist: Fix unstable selection of packages with the same version
If there are multiple ebuilds with the same version from different
repositories, we need to select the ebuild in the repository with the
highest priority. This change fixes the unstable iteration order and
unstable sort so that we properly select the correct ebuild.

The sort_by_key and reverse in dump_package is only there to make the
output easier to understand.

BUG=b:269606353
TEST=cargo run -- --board arm64-generic generate-repo --output-dir ~/tmp/board-stable/
Perform generation multiple times and verified output is stable.

Change-Id: Ieb2fbdcc699dcd9d0a50d82684f04334fe4ba5ad
Reviewed-on: https://team-review.git.corp.google.com/c/cros-build-tiger/cros-bazel/+/1753820
Reviewed-by: Ryo Hashimoto <hashimoto@google.com>
Reviewed-by: Shuhei Takahashi <nya@google.com>